move levelgenerator into contrib
