+++ game/src/contrib/level/generator/GeneratorUtils.java
@@ -1,9 +1,12 @@
-package core.level.utils;
+package contrib.level.generator;
+
+import contrib.level.generator.graphBased.levelGraph.Direction;
 
 import core.level.Tile;
 import core.level.elements.ILevel;
 import core.level.elements.tile.DoorTile;
-import core.level.generator.graphBased.levelGraph.Direction;
+import core.level.utils.LevelElement;
+import core.level.utils.TileTextureFactory;
 
 import java.util.Optional;
