+++ game/src/contrib/level/generator/perlinNoise/PerlinNoiseGenerator.java
@@ -1,4 +1,4 @@
-package core.level.generator.perlinNoise;
+package contrib.level.generator.perlinNoise;
 
 import core.level.TileLevel;
 import core.level.elements.ILevel;
